requirements.txt
amqp==5.1.1
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
Django==4.2.6
django-filter==23.3
django-redis==5.4.0
django-shell-plus==1.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
executing==2.0.1
flower==2.0.1
humanize==4.8.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.17.2
jedi==0.19.1
kombu==5.3.2
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pluggy==1.3.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
pytest==7.4.3
pytest-django==4.6.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.9
